Add example for user-onboarding
The workflow is an example workflow for testing purposes, for testing
access to external systems, loops and sleep.
The workflow registers a new user into a system, and after 3 checks it
marks the user onboarding as completed.
Between cycles the workflows sleeps for few seconds.

The example contains the workflow itself and also the user-onboarding
server and its spec files.

Signed-off-by: Moti Asayag <masayag@redhat.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED